Add Storage and Ui classes
