+++ src/data/sirtalksalot.txt
@@ -1,5 +1,2 @@
-[T] [-] dfsdfsdf 
-[D] [-] asdfdsa (by: 01/12/2023)
-[D] [X] dadfas (by: 07/03/2025)
-[D] [-] dfasdf (by: 03/09/3038)
-[E] [-] jhjkjkhhk (from: 01/02/2023 to: 04/02/2023)
+[T] [-] fasdf 
+[E] [-] dfasldfjdsalf (from: 04/05/2003 to: 02/04/2004)
