+++ src/main/java/Parser.java
@@ -1,0 +1,2 @@
+class Parser {
+}
