+++ src/main/java/SirTalksALot.java
@@ -1,3 +1,5 @@
+import com.sun.source.util.TaskListener;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
@@ -7,7 +9,6 @@ import java.util.Arrays;
 import java.util.Scanner;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.time.temporal.ChronoUnit;
 
 class Task {
     private String name;
@@ -114,24 +115,33 @@ class Event extends Task {
 }
 
 public class SirTalksALot {
-    public static void main(String[] args) {
+    private Ui ui;
+    private Storage storage;
+    private Parser parser;
+    private TaskList taskList;
+
+    public SirTalksALot(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        taskList = new TaskList(storage.loadData());
+    }
+
+    public void run() {
         Scanner scanner = new Scanner(System.in);
-        ArrayList<Task> taskList;
+        ArrayList<Task> taskList = new ArrayList<>();
 
-        taskList = loadData();
-
-        System.out.println("____________________________________________________________");
-        sayHello();
-        System.out.println("____________________________________________________________");
+        Ui.breakLine();
+        Ui.sayHello();
+        Ui.breakLine();
 
         String[] input = scanner.nextLine().split(" ");
         while (!input[0].equals("bye")) {
-            System.out.println("____________________________________________________________");
+            Ui.breakLine();
 
             switch (input[0]) {
             case "list" -> {
                 if (taskList.isEmpty()) {
-                    countTask(0);
+                    Ui.countTask(0);
                     break;
                 }
                 System.out.println("Hear ye! These are the tasks upon thy list, as decreed by thine own hand:");
@@ -151,8 +161,8 @@ public class SirTalksALot {
                     System.out.println("Noted, then. I have seen fit to remove this trivial task.");
                     System.out.println(taskList.get(index).getType() + taskList.get(index).isComplete() + " " + taskList.get(index).getName());
                     taskList.remove(index);
-                    saveData(taskList);
-                    countTask(taskList.size());
+                    storage.saveData(taskList);
+                    Ui.countTask(taskList.size());
                 } catch (IndexOutOfBoundsException e) {
                     System.out.println("It seems no such task exists, for I have heard no mention of it, nor does it appear to fall within the realm of possibility.");
                 }
@@ -161,7 +171,7 @@ public class SirTalksALot {
                 try {
                     int index = Integer.parseInt(input[1]) - 1;
                     taskList.get(index).markCompleted();
-                    saveData(taskList);
+                    storage.saveData(taskList);
                     System.out.println("Behold! A task completed! A most noble accomplishment.");
                     System.out.println(taskList.get(index).getType() + "[X] " + taskList.get(index).getName());
                 } catch (IndexOutOfBoundsException e) {
@@ -172,7 +182,7 @@ public class SirTalksALot {
                 try {
                     int index = Integer.parseInt(input[1]) - 1;
                     taskList.get(index).unmarkCompleted();
-                    saveData(taskList);
+                    storage.saveData(taskList);
                     System.out.println("Alas, a task left to be conquered. Its time has not yet come to pass.");
                     System.out.println(taskList.get(index).getType() + "[ ] " + taskList.get(index).getName());
                 } catch (IndexOutOfBoundsException e) {
@@ -186,11 +196,11 @@ public class SirTalksALot {
                 } else {
                     String todo = String.join(" ", Arrays.copyOfRange(input, 1, input.length));
 
-                    addTask();
+                    Ui.addTask();
                     taskList.add(new ToDo(todo));
-                    saveData(taskList);
+                    storage.saveData(taskList);
                     System.out.println("    [T][-] " + taskList.get(taskList.size() - 1).getName());
-                    countTask(taskList.size());
+                    Ui.countTask(taskList.size());
                 }
             }
             case "deadline" -> {
@@ -209,12 +219,12 @@ public class SirTalksALot {
                     break;
                 }
 
-                addTask();
+                Ui.addTask();
                 taskList.add(new DeadLine(temp[0], LocalDate.parse(temp[1])));
-                saveData(taskList);
+                storage.saveData(taskList);
                 System.out.println("    [D][-] " + taskList.get(taskList.size() - 1).getName()
                         + " (by: " + taskList.get(taskList.size() - 1).getDeadLine() + ")");
-                countTask(taskList.size());
+                Ui.countTask(taskList.size());
             }
             case "event" -> {
                 if (input.length == 1) {
@@ -238,132 +248,27 @@ public class SirTalksALot {
                     break;
                 }
 
-                addTask();
+                Ui.addTask();
                 taskList.add(new Event(temp[0], LocalDate.parse(temp1[0]), LocalDate.parse(temp1[1])));
-                saveData(taskList);
+                storage.saveData(taskList);
                 System.out.println("    [E][-] " + taskList.get(taskList.size() - 1).getName() + " (from: " + temp1[0] + " to: " + temp1[1] + ")");
-                countTask(taskList.size());
+                Ui.countTask(taskList.size());
             }
             default -> {
                 System.out.println("Speak clearly! I know not what such words mean. Attempt once more.");
             }
             }
-            System.out.println("____________________________________________________________");
+            Ui.breakLine();
             input = scanner.nextLine().split(" ");
         }
 
-        System.out.println("____________________________________________________________");
-        sayBye();
-        saveData(taskList);
-        System.out.println("____________________________________________________________");
+        Ui.breakLine();
+        Ui.sayBye();
+        storage.saveData(taskList);
+        Ui.breakLine();
     }
 
-    public static void saveData(ArrayList<? extends Task> taskList) {
-        try {
-            String directory = "data";
-            String filename = directory + File.separator + "sirtalksalot.txt";
-            File d = new File(directory);
-            File f = new File(filename);
-
-            if (!f.exists()) {
-                d.mkdirs();
-                f.createNewFile();
-            }
-
-            FileWriter fw = new FileWriter(filename);
-            for (Task task : taskList) {
-                String type = task.getType();
-                fw.write(type + " " + task.isComplete() + " " + task.getName() + " " + task.toString() + "\n");
-            }
-            fw.close();
-
-        } catch (IOException e) {
-            System.out.println(e.getMessage());
-        }
-    }
-
-    public static ArrayList<Task> loadData() {
-        ArrayList<Task> taskList = new ArrayList<>();
-        try {
-            String directory = "data";
-            String filename = directory + File.separator + "sirtalksalot.txt";
-            File f = new File(filename);
-            if (f.exists()) {
-                Scanner sc = new Scanner(f);
-                while (sc.hasNextLine()) {
-                    String[] task = sc.nextLine().split(" ");
-                    switch (task[0]) {
-                    case "[T]" -> {
-                        ToDo toDo = new ToDo(task[2]);
-                        if (task[1].equals("[X]")) {
-                            toDo.markCompleted();
-                        }
-                        taskList.add(toDo);
-                    }
-                    case "[D]" -> {
-                        String token = String.join(" ", Arrays.copyOfRange(task, 2, task.length));
-                        String[] token1 = token.split(" \\(by: ", 2);
-                        DeadLine deadLine = new DeadLine(token1[0],
-                                LocalDate.parse(token1[1].substring(0, token1[1].length() - 1),
-                                        DateTimeFormatter.ofPattern("dd/MM/yyyy")));
-                        if (task[1].equals("[X]")) {
-                            deadLine.markCompleted();
-                        }
-                        taskList.add(deadLine);
-                    }
-                    case "[E]" -> {
-                        String token = String.join(" ", Arrays.copyOfRange(task, 2, task.length));
-                        String[] token1 = token.split(" \\(from: ", 2);
-                        String[] token2 = token1[1].split(" to: ", 2);
-                        Event event = new Event(token1[0],
-                                LocalDate.parse(token2[0], DateTimeFormatter.ofPattern("dd/MM/yyyy")),
-                                LocalDate.parse(token2[1].substring(0, token2[1].length() - 1),
-                                        DateTimeFormatter.ofPattern("dd/MM/yyyy")));
-                        if (task[1].equals("[X]")) {
-                            event.markCompleted();
-                        }
-                        taskList.add(event);
-                    }
-                    default -> System.out.println("Wrong file format!");
-                    }
-                }
-            }
-        } catch (FileNotFoundException e) {
-            System.out.println(e.getMessage());
-        }
-        return taskList;
-    }
-
-    public static void addTask() {
-        String addTask = "Verily, I have inscribed this task upon the list.\n"
-                + "Let no task go unfulfilled and no duty unrecorded!";
-        System.out.println(addTask);
-    }
-
-    public static void countTask(int count) {
-        String countTask = "";
-        if (count == 0) {
-            countTask = "Thou hast " + count + " tasks upon the list.";
-        } else if (count == 1) {
-            countTask = "Thou hast " + count + " task upon the list. A worthy pursuit!";
-        } else {
-            countTask = "Thou hast " + count + " tasks upon the list, each one a worthy pursuit!";
-        }
-        System.out.println(countTask);
-    }
-
-    public static void sayHello() {
-        String greeting = "Greetings! It is I, Sir Talks-A-Lot. \n"
-                + "Ah, so you wish to converse with one such as myself? \n"
-                + "A noble knight, sworn to honor and valor, holder of great wisdom and unyielding strength? \n"
-                + "Very well, I shall indulge thee in thy request. \n"
-                + "Speak now, peasant, What dost thou seek from a knight of my stature?";
-        System.out.println(greeting);
-    }
-
-    public static void sayBye() {
-        String bye = "Hah! A quick departure, is it? Very well, then. \n"
-                + "Farewell for now.";
-        System.out.println(bye);
+    public static void main(String[] args) {
+        new SirTalksALot("data/sirtalksalot.txt").run();
     }
 }
