+++ src/main/java/TaskList.java
@@ -1,0 +1,9 @@
+import java.util.ArrayList;
+
+class TaskList {
+    ArrayList<Task> taskList;
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.taskList = tasks;
+    }
+}
